+++ lib/active_admin/object_mappers/active_record.rb
@@ -1,0 +1,3 @@
+# ActiveRecord-specific plugins should be required here
+
+require 'active_admin/object_mappers/active_record/comments'
