+++ lib/active_admin/object_mappers/active_record/comments/comment.rb
