+++ lib/active_admin/object_mappers/active_record/comments/views.rb
@@ -1,0 +1,2 @@
+require 'active_admin/views'
+require 'active_admin/object_mappers/active_record/comments/views/active_admin_comments'
