+++ lib/active_admin/object_mappers/mongoid.rb
