--- lib/active_admin/orm/active_record.rb
@@ -1,3 +1,0 @@
-# ActiveRecord-specific plugins should be required here
-
-require 'active_admin/orm/active_record/comments'
