--- lib/active_admin/orm/active_record/comments/views.rb
@@ -1,2 +1,0 @@
-require 'active_admin/views'
-require 'active_admin/orm/active_record/comments/views/active_admin_comments'
